requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2==2.7.6.1
python-dotenv==0.9.1
six==1.11.0
Werkzeug==0.14.1
